refactor: 服务列表UI重构

+++ src/pages/ServiceList/serviceListByType.tsx
@@ -1,18 +1,22 @@
 import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import { useHistory, useParams } from 'react-router-dom';
-import {
-	Button,
-	Message,
-	Dialog,
-	Checkbox,
-	Balloon,
-	Loading
-} from '@alicloud/console-components';
-import { Page, Content, Header } from '@alicloud/console-components-page';
-import Actions, { LinkButton } from '@alicloud/console-components-actions';
+// import {
+// 	Button,
+// 	Message,
+// 	Dialog,
+// 	Checkbox,
+// 	Balloon,
+// 	Loading
+// } from '@alicloud/console-components';
+import { Button, notification, Modal, Checkbox, Tooltip } from 'antd';
+// import { Page, Content, Header } from '@alicloud/console-components-page';
+// import Actions, { LinkButton } from '@alicloud/console-components-actions';
 // --- 自定义组件
-import Table from '@/components/MidTable';
+import { ProPage, ProContent, ProHeader } from '@/components/ProPage';
+import ProTable from '@/components/ProTable';
+import Actions from '@/components/Actions';
+// import Table from '@/components/MidTable';
 // --- 方法
 import {
 	getList,
@@ -45,8 +49,11 @@ import storage from '@/utils/storage';
 import { states } from '@/utils/const';
 import { serviceListStatusRender, timeRender, nullRender } from '@/utils/utils';
 import GuidePage from '../GuidePage';
+import { CheckboxChangeEvent } from 'antd/lib/checkbox';
 // --- css样式
 
+const { confirm } = Modal;
+const LinkButton = Actions.LinkButton;
 const tabJudge: (record: serviceProps, tab: string) => boolean = (
 	record: serviceProps,
 	tab: string
@@ -61,7 +68,7 @@ const tabJudge: (record: serviceProps, tab: string) => boolean = (
 		}
 	}
 };
-const Tooltip = Balloon.Tooltip;
+// const Tooltip = Balloon.Tooltip;
 
 const ServiceListByType = (props: serviceListProps) => {
 	const { setCluster, setNamespace, setRefreshCluster } = props;
@@ -81,17 +88,10 @@ const ServiceListByType = (props: serviceListProps) => {
 	const history = useHistory();
 	const params: ListParamsProps = useParams();
 	const { name, aliasName } = params;
-	const [lock, setLock] = useState<any>({ lock: 'right' });
+	// const [lock, setLock] = useState<any>({ lock: 'right' });
 	const [middlewareInfo, setMiddlewareInfo] = useState<middlewareProps>();
 	const [loadingVisible, setLoadingVisible] = useState<boolean>(true);
 
-	useEffect(() => {
-		window.onresize = function () {
-			document.body.clientWidth >= 2300
-				? setLock(null)
-				: setLock({ lock: 'right' });
-		};
-	}, []);
 	useEffect(() => {
 		let mounted = true;
 		if (JSON.stringify(namespace) !== '{}') {
@@ -114,7 +114,10 @@ const ServiceListByType = (props: serviceListProps) => {
 								setShowDataSource([]);
 							}
 						} else {
-							Message.show(messageConfig('error', '失败', res));
+							notification.error({
+								message: '失败',
+								description: res.errorMsg
+							});
 						}
 					})
 					.finally(() => {
@@ -132,7 +135,10 @@ const ServiceListByType = (props: serviceListProps) => {
 							setCantRelease(true);
 						}
 					} else {
-						Message.show(messageConfig('error', '失败', res));
+						notification.error({
+							message: '失败',
+							description: res.errorMsg
+						});
 					}
 				});
 				getCanReleaseMiddleware({
@@ -142,7 +148,10 @@ const ServiceListByType = (props: serviceListProps) => {
 					if (res.success) {
 						setMiddlewareInfo(res.data);
 					} else {
-						Message.show(messageConfig('error', '失败', res));
+						notification.error({
+							message: '失败',
+							description: res.errorMsg
+						});
 					}
 				});
 			}
@@ -171,21 +180,24 @@ const ServiceListByType = (props: serviceListProps) => {
 						setShowDataSource([]);
 					}
 				} else {
-					Message.show(messageConfig('error', '失败', res));
+					notification.error({
+						message: '失败',
+						description: res.errorMsg
+					});
 				}
 			})
 			.finally(() => {
 				setLoadingVisible(false);
 			});
 	};
-	const handleChange: (value: string) => void = (value: string) => {
-		setKeyword(value);
+	const handleChange: (e: any) => void = (e: any) => {
+		setKeyword(e.target.value);
 	};
 	const handleSearch = () => {
 		getData();
 	};
 	const deleteFn = (record: serviceProps) => {
-		Dialog.show({
+		confirm({
 			title: '提示',
 			content: '确定删除该服务？',
 			onOk: () => {
@@ -197,11 +209,15 @@ const ServiceListByType = (props: serviceListProps) => {
 				})
 					.then((res) => {
 						if (res.success) {
-							Message.show(
-								messageConfig('success', '成功', '删除成功')
-							);
+							notification.success({
+								message: '成功',
+								description: '删除成功'
+							});
 						} else {
-							Message.show(messageConfig('error', '失败', res));
+							notification.error({
+								message: '失败',
+								description: res.errorMsg
+							});
 						}
 					})
 					.finally(() => {
@@ -258,8 +274,8 @@ const ServiceListByType = (props: serviceListProps) => {
 			return { style: { background: '#F8F8F9', color: '#CCCCCC' } };
 		}
 	};
-	const handleFilterBackup = (checked: boolean) => {
-		setBackupCheck(checked);
+	const handleFilterBackup = (e: CheckboxChangeEvent) => {
+		setBackupCheck(e.target.value);
 		let list = dataSource?.serviceList || [];
 		if (selectedKeys.length > 0) {
 			if (selectedKeys[0] !== 'Other') {
@@ -274,7 +290,7 @@ const ServiceListByType = (props: serviceListProps) => {
 				});
 			}
 		}
-		if (checked) {
+		if (e.target.value) {
 			list = showDataSource.filter(
 				(item) => item?.mysqlDTO?.openDisasterRecoveryMode === true
 			);
@@ -331,22 +347,22 @@ const ServiceListByType = (props: serviceListProps) => {
 			middlewareName: record.name,
 			type: record.type
 		};
-		Dialog.show({
+		confirm({
 			title: '操作确认',
 			content: '请确认是否恢复该服务！',
 			onOk: () => {
 				return recoveryMiddleware(sendData)
 					.then((res) => {
 						if (res.success) {
-							Message.show(
-								messageConfig(
-									'success',
-									'成功',
-									'该服务已恢复,3秒后刷新'
-								)
-							);
+							notification.success({
+								message: '成功',
+								description: '该服务已恢复,3秒后刷新'
+							});
 						} else {
-							Message.show(messageConfig('error', '失败', res));
+							notification.error({
+								message: '失败',
+								description: res.errorMsg
+							});
 						}
 					})
 					.finally(() => {
@@ -364,22 +380,22 @@ const ServiceListByType = (props: serviceListProps) => {
 			middlewareName: record.name,
 			type: record.type
 		};
-		Dialog.show({
+		confirm({
 			title: '操作确认',
 			content: '删除后无法恢复该服务，请谨慎操作！',
 			onOk: () => {
 				return deleteMiddlewareStorage(sendData)
 					.then((res) => {
 						if (res.success) {
-							Message.show(
-								messageConfig(
-									'success',
-									'成功',
-									'该服务已彻底删除'
-								)
-							);
+							notification.success({
+								message: '成功',
+								description: '该服务已彻底删除'
+							});
 						} else {
-							Message.show(messageConfig('error', '失败', res));
+							notification.error({
+								message: '失败',
+								description: res.errorMsg
+							});
 						}
 					})
 					.finally(() => {
@@ -414,17 +430,13 @@ const ServiceListByType = (props: serviceListProps) => {
 			if (name === 'mysql') {
 				return {
 					primary: (
-						<Tooltip
-							trigger={
-								<Button
-									type="primary"
-									disabled={!createFlag || !getFlag}
-								>
-									发布服务
-								</Button>
-							}
-						>
-							当前用户无该操作的权限!
+						<Tooltip title="当前用户无该操作的权限!">
+							<Button
+								type="primary"
+								disabled={!createFlag || !getFlag}
+							>
+								发布服务
+							</Button>
 						</Tooltip>
 					),
 					secondary: (
@@ -439,17 +451,13 @@ const ServiceListByType = (props: serviceListProps) => {
 			} else {
 				return {
 					primary: (
-						<Tooltip
-							trigger={
-								<Button
-									type="primary"
-									disabled={!createFlag || !getFlag}
-								>
-									发布服务
-								</Button>
-							}
-						>
-							当前用户无该操作的权限!
+						<Tooltip title="当前用户无该操作的权限!">
+							<Button
+								type="primary"
+								disabled={!createFlag || !getFlag}
+							>
+								发布服务
+							</Button>
 						</Tooltip>
 					)
 				};
@@ -459,18 +467,14 @@ const ServiceListByType = (props: serviceListProps) => {
 			if (name === 'mysql') {
 				return {
 					primary: (
-						<Tooltip
-							trigger={
-								<Button
-									onClick={releaseMiddleware}
-									type="primary"
-									disabled={cantRelease}
-								>
-									发布服务
-								</Button>
-							}
-						>
-							请前往平台组件界面安装中间件管理组件！
+						<Tooltip title="请前往平台组件界面安装中间件管理组件！">
+							<Button
+								onClick={releaseMiddleware}
+								type="primary"
+								disabled={cantRelease}
+							>
+								发布服务
+							</Button>
 						</Tooltip>
 					),
 					secondary: (
@@ -485,18 +489,14 @@ const ServiceListByType = (props: serviceListProps) => {
 			} else {
 				return {
 					primary: (
-						<Tooltip
-							trigger={
-								<Button
-									onClick={releaseMiddleware}
-									type="primary"
-									disabled={cantRelease}
-								>
-									发布服务
-								</Button>
-							}
-						>
-							请前往平台组件界面安装中间件管理组件！
+						<Tooltip title="请前往平台组件界面安装中间件管理组件！">
+							<Button
+								onClick={releaseMiddleware}
+								type="primary"
+								disabled={cantRelease}
+							>
+								发布服务
+							</Button>
 						</Tooltip>
 					)
 				};
@@ -505,18 +505,14 @@ const ServiceListByType = (props: serviceListProps) => {
 			if (name === 'mysql') {
 				return {
 					primary: (
-						<Tooltip
-							trigger={
-								<Button
-									onClick={releaseMiddleware}
-									type="primary"
-									disabled={!middlewareInfo}
-								>
-									发布服务
-								</Button>
-							}
-						>
-							数据加载中，请稍后...
+						<Tooltip title="数据加载中，请稍后...">
+							<Button
+								onClick={releaseMiddleware}
+								type="primary"
+								disabled={!middlewareInfo}
+							>
+								发布服务
+							</Button>
 						</Tooltip>
 					),
 					secondary: (
@@ -531,18 +527,14 @@ const ServiceListByType = (props: serviceListProps) => {
 			} else {
 				return {
 					primary: (
-						<Tooltip
-							trigger={
-								<Button
-									onClick={releaseMiddleware}
-									type="primary"
-									disabled={!middlewareInfo}
-								>
-									发布服务
-								</Button>
-							}
-						>
-							数据加载中，请稍后...
+						<Tooltip title="数据加载中，请稍后...">
+							<Button
+								onClick={releaseMiddleware}
+								type="primary"
+								disabled={!middlewareInfo}
+							>
+								发布服务
+							</Button>
 						</Tooltip>
 					)
 				};
@@ -585,8 +577,8 @@ const ServiceListByType = (props: serviceListProps) => {
 	};
 	const actionRender = (
 		value: string,
-		index: number,
-		record: serviceProps
+		record: serviceProps,
+		index: number
 	) => {
 		const jsonRole: User = JSON.parse(storage.getLocal('role'));
 		console.log(jsonRole);
@@ -667,18 +659,16 @@ const ServiceListByType = (props: serviceListProps) => {
 											: record.type === 'rocketmq'
 											? `${record.name}-console-svc`
 											: `${record.name}-manager-svc`;
-									Message.show(
-										messageConfig(
-											'error',
-											'失败',
-											`请先前往“服务暴露”暴露该服务的${sn}服务`
-										)
-									);
+									notification.error({
+										message: '失败',
+										description: `请先前往“服务暴露”暴露该服务的${sn}服务`
+									});
 								}
 							} else {
-								Message.show(
-									messageConfig('error', '失败', res)
-								);
+								notification.error({
+									message: '失败',
+									description: res.errorMsg
+								});
 							}
 						});
 					}}
@@ -712,9 +702,10 @@ const ServiceListByType = (props: serviceListProps) => {
 	};
 	const toDetail = (record: any) => {
 		if (!record.mysqlDTO.relationExist) {
-			Message.show(
-				messageConfig('error', '失败', '该关联实例不存在，无法进行跳转')
-			);
+			notification.error({
+				message: '失败',
+				description: '该关联实例不存在，无法进行跳转'
+			});
 			return;
 		}
 		const cs = globalClusterList.filter(
@@ -741,8 +732,8 @@ const ServiceListByType = (props: serviceListProps) => {
 	};
 	const associatedRender = (
 		value: string,
-		index: number,
-		record: serviceProps
+		record: serviceProps,
+		index: number
 	) => {
 		return (
 			<div className="display-flex flex-align">
@@ -788,7 +779,7 @@ const ServiceListByType = (props: serviceListProps) => {
 			</div>
 		);
 	};
-	const nameRender = (value: string, index: number, record: serviceProps) => {
+	const nameRender = (value: string, record: serviceProps, index: number) => {
 		if (record.status === 'Deleted') {
 			return (
 				<div style={{ maxWidth: '160px' }}>
@@ -871,7 +862,7 @@ const ServiceListByType = (props: serviceListProps) => {
 			</div>
 		);
 	};
-	const podRender = (value: string, index: number, record: serviceProps) => {
+	const podRender = (value: string, record: serviceProps, index: number) => {
 		const jsonRole: User = JSON.parse(storage.getLocal('role'));
 		let operateFlag = false;
 		if (jsonRole.userRoleList.some((i: any) => i.roleId === 1)) {
@@ -904,21 +895,21 @@ const ServiceListByType = (props: serviceListProps) => {
 		return <GuidePage />;
 	}
 	return (
-		<Page>
-			<Header
+		<ProPage>
+			<ProHeader
 				title={`${aliasName || ''}服务列表`}
 				subTitle="已发布中间件服务管理列表"
 			/>
-			<Content>
-				<Table
+			<ProContent>
+				<ProTable
 					dataSource={showDataSource}
-					exact
-					fixedBarExpandWidth={[24]}
-					affixActionBar
+					// exact
+					// fixedBarExpandWidth={[24]}
+					// affixActionBar
 					showColumnSetting
 					showRefresh
 					onRefresh={getData}
-					primaryKey="name"
+					rowKey="name"
 					operation={operation()}
 					loading={loadingVisible}
 					search={{
@@ -927,59 +918,59 @@ const ServiceListByType = (props: serviceListProps) => {
 						onSearch: handleSearch,
 						placeholder: '请输入搜索内容'
 					}}
-					onSort={onSort}
-					onFilter={onFilter}
-					rowProps={onRowProps}
+					// onSort={onSort}
+					// onFilter={onFilter}
+					// onRow={onRowProps}
 				>
-					<Table.Column
+					<ProTable.Column
 						title="服务名称/中文别名"
 						dataIndex="name"
 						width={180}
-						cell={nameRender}
-						lock="left"
+						render={nameRender}
+						fixed="left"
 					/>
-					<Table.Column
+					<ProTable.Column
 						title="状态"
 						dataIndex="status"
 						width={150}
-						cell={serviceListStatusRender}
-						filters={states}
-						filterMode="single"
+						render={serviceListStatusRender}
+						// filters={states}
+						// filterMode="single"
 					/>
-					<Table.Column
+					<ProTable.Column
 						title="实例数"
 						dataIndex="podNum"
-						cell={podRender}
+						render={podRender}
 						width={80}
 					/>
-					<Table.Column
+					<ProTable.Column
 						title="备注"
 						dataIndex="description"
-						cell={nullRender}
+						render={nullRender}
 					/>
-					<Table.Column
+					<ProTable.Column
 						title="关联服务名称/中文别名"
 						dataIndex="associated"
 						width={180}
-						cell={associatedRender}
+						render={associatedRender}
 					/>
-					<Table.Column
+					<ProTable.Column
 						title="创建时间"
 						dataIndex="createTime"
 						width={180}
-						sortable={true}
-						cell={timeRender}
+						// sortable={true}
+						render={timeRender}
 					/>
-					<Table.Column
+					<ProTable.Column
 						title="操作"
 						dataIndex="action"
-						cell={actionRender}
+						render={actionRender}
 						width={300}
-						{...lock}
+						// {...lock}
 					/>
-				</Table>
-			</Content>
-		</Page>
+				</ProTable>
+			</ProContent>
+		</ProPage>
 	);
 };
 const mapStateToProps = (state: StoreState) => ({
